% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_exporting.R
\name{as.data.frame.FeatureTable}
\alias{as.data.frame.FeatureTable}
\title{Return FeatureTable data field as a data.frame.}
\usage{
\method{as.data.frame}{FeatureTable}(ft)
}
\arguments{
\item{ft}{A FeatureTable}
}
\value{
The \code{data} field of the \code{FeatureTable} as a \code{data.frame}.
}
\description{
Return FeatureTable data field as a data.frame.
}
