% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{clr}
\alias{clr}
\title{Centered log ratio.}
\usage{
clr(ft, base = 2)
}
\arguments{
\item{ft}{A Feature table.}

\item{base}{Base of logarithm.}
}
\value{
A FeatureTable with the \code{data} transformed with centered log ratio.
}
\description{
Centered log ratio.
}
