% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{core_microbiome.FeatureTable}
\alias{core_microbiome.FeatureTable}
\title{Core Microbiome}
\usage{
\method{core_microbiome}{FeatureTable}(
  ft,
  detection_limit = 1,
  min_sample_proportion = NULL,
  max_sample_proportion = NULL,
  min_samples = NULL,
  max_samples = NULL
)
}
\arguments{
\item{ft}{A FeatureTable}

\item{detection_limit}{Limit of detection for a feature to be considered
present in a sample}

\item{min_sample_proportion}{Minimum proportion of samples in which a feature
must be present to be kept.}

\item{max_sample_proportion}{Maximum proportion of samples in which a feature
must be present to be kept.}

\item{min_samples}{Minimum number of samples in which a feature must be
present to be kept.}

\item{max_samples}{Maximum number of samples in which a feature must be
present to be kept.}
}
\value{
A new FeatureTable with only the specified "core" features.
}
\description{
Core Microbiome
}
\details{
If both \code{min_sample_proportion} and \code{min_samples} are given, an
  error will be raised. If \code{min_sample_proportion} is not a proportion,
  an error will be raised.  If \code{min_samples} looks like a proportion, a
   warning will be given.
}
