% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{dim.FeatureTable}
\alias{dim.FeatureTable}
\title{Dimensions of a FeatureTable}
\usage{
\method{dim}{FeatureTable}(ft)
}
\arguments{
\item{ft}{A FeatureTable object.}
}
\value{
The dimension of the feature_table.
}
\description{
Return the dimensions of the FeatureTable.
}
\seealso{
\code{\link{dim}} for generic method, \code{\link{FeatureTable}} for R6 method.

Other dimension functions: 
\code{\link{nfeatures}()},
\code{\link{nobservations}()},
\code{\link{nsamples}()},
\code{\link{num_features}()},
\code{\link{num_observations}()},
\code{\link{num_samples}()}
}
\concept{dimension functions}
