% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{feature_names}
\alias{feature_names}
\alias{feature_names.FeatureTable}
\title{Feature/column names of a FeatureTable.}
\usage{
feature_names(ft)

\method{feature_names}{FeatureTable}(ft)
}
\arguments{
\item{ft}{A FeatureTable object.}
}
\value{
The names of samples/observations/rows in the FeatureTable.
}
\description{
Return the names of samples/observations/rows in the FeatureTable.
}
\seealso{
Other row name functions: 
\code{\link{sample_names}()}
}
\concept{row name functions}
