% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft}
\alias{ft}
\title{A FeatureTable with count-like data.}
\format{
A FeatureTable
\describe{
  \item{feature_table}{4 sample X 5 feature table with count-like data}
  \item{feature_data}{3 variables, \code{Color}, \code{Shape}, and \code{Length}}
  \item{sample_data}{3 variables, \code{Location}, \code{Season}, and \code{SnazzyFactor}}
}
}
\usage{
ft
}
\description{
A small FeatureTable with count-like (e.g., OTU-like) data used for
examples and to test FeatureTable functions.
}
\keyword{datasets}
