% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_exporting.R
\name{ft_from.phyloseq}
\alias{ft_from.phyloseq}
\title{Create a FeatureTable from a phyloseq object.}
\usage{
\method{ft_from}{phyloseq}(x, numeric_feature_data_columns = NULL, ...)
}
\arguments{
\item{x}{A \code{phyloseq} object.}

\item{numeric_feature_data_columns}{A vector specifying which of the data
columns are numeric.  See details.}
}
\value{
A FeatureTable
}
\description{
Create a FeatureTable from a phyloseq object.
}
\details{
The \code{phyloseq::otu_table} becomes the \code{data}.

The \code{phyloseq::tax_table} becomes the \code{feature_data}.

The \code{phyloseq::sample_data} becomes the \code{sample_data}.

Normally from a phyloseq object, the tax table wouldn't have numeric values.
BUT if it comes from FT => phy => FT it could.  Since phyloseq taxtable does
NOT play well with numbers, the user must specify any numeric
feature_data columns manually :(
}
