% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{is.FeatureTable}
\alias{is.FeatureTable}
\title{Is an Object a FeatureTable?}
\usage{
is.FeatureTable(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if Object is a \code{FeatureTable}, \code{FALSE} otherwise.
}
\description{
Is an Object a FeatureTable?
}
