% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{is_count_table}
\alias{is_count_table}
\alias{is_count_table.FeatureTable}
\title{Is the FeatureTable a count table?}
\usage{
is_count_table(ft)

\method{is_count_table}{FeatureTable}(ft)
}
\arguments{
\item{ft}{A FeatureTable}
}
\value{
TRUE if \code{feature_table} contains only natural numbers (counting numbers >= 0), FALSE otherwise.
}
\description{
Is the FeatureTable a count table?
}
