% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{keep}
\alias{keep}
\title{Keep samples/observations/rows using a predicate}
\usage{
keep(ft, ...)
}
\arguments{
\item{ft}{A FeatureTable object.}

\item{...}{Optional arguments to \code{predicate} (if \code{predicate} is a function).}

\item{margin}{Subscript which the predicate will be applied over.  E.g., \code{1} or \code{"samples"} indicates rows, \code{2} or \code{"features"} indicates columns.}

\item{predicate}{The predicate function or expression to be applied.  Only those elements where \code{predicate} is or evaluates to \code{TRUE} will be kept.}
}
\value{
A new FeatureTable with the elements that were kept.
}
\description{
Keep samples/observations/rows using a predicate function or expression.
}
\seealso{
Other filtering functions: 
\code{\link{keep_features}()},
\code{\link{keep_samples}()}
}
\concept{filtering functions}
