% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{keep_samples}
\alias{keep_samples}
\title{Keep elements using a predicate}
\usage{
keep_samples(ft, predicate, ...)
}
\arguments{
\item{ft}{A FeatureTable object.}

\item{predicate}{The predicate function or expression to be applied.  Only those elements where \code{predicate} is or evaluates to \code{TRUE} will be kept.}

\item{...}{Optional arguments to \code{predicate} (if \code{predicate} is a function).}
}
\value{
A new FeatureTable with the elements that were kept.
}
\description{
Keep elements using a predicate function or expression.
}
\seealso{
Other filtering functions: 
\code{\link{keep_features}()},
\code{\link{keep}()}
}
\concept{filtering functions}
