% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee}
\alias{lee}
\title{A FeatureTable with data from Lee et al. (2015)}
\format{
A FeatureTable
\describe{
  \item{feature_table}{16 samples X 1490 features table with 16S abundance}
  \item{feature_data}{7 variables, taxonomy}
  \item{sample_data}{4 variables, \code{Temp}, \code{Type}, \code{Char}, \code{Color}}
}
}
\usage{
# Import into the global env.
data(lee)

# Refer to the FeatureTable directly.
featuretable::lee
}
\description{
A FeatureTable with count table, sample data, and taxonomy of an SSU rRNA
study of basalts on the Dorado Outcrop.
}
\keyword{datasets}
