% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee_phyloseq}
\alias{lee_phyloseq}
\title{Lee data as a phyloseq object.}
\format{
An object of class \code{phyloseq} of length 1.
}
\usage{
lee_phyloseq
}
\description{
Mainly for testing.
}
\keyword{datasets}
