% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{map_features}
\alias{map_samples}
\title{Apply a function over FeatureTable margins and return a new FeatureTable.}
\usage{
map(ft, margin, fn, ...)

map_features(ft, fn, ...)

map_samples(ft, fn, ...)
}
\arguments{
\item{ft}{A FeatureTable.}

\item{margin}{Margin to apply the function over.  E.g., \code{1} or
\code{"samples"} indicates rows, \code{2} or \code{"features"} indicates
columns.}

\item{fn}{The function to be applied over the specified margin.}

\item{...}{Optional arguments to \code{fn}.}
}
\value{
A new FeatureTable with the result of applying \code{fn} over
  \code{margin}.
}
\description{
Returns a new FeatureTable by applying a function to margins of the
\code{$data} field of a \code{FeatureTable}.
}
\details{
Named helpers can also be used rather than specifying the \code{margin}
argument.

\itemize{
  \item \code{\link{map_features}}
  \item \code{\link{map_samples}}
}
}
\examples{
data(ft)

## Using R6 methods

# Map over features
ft$map("features", function(feature) feature / sum(feature))

# Map over samples
ft$map("samples", function(sample) sample / sum(sample))

## Using s3 methods

# Map over features
map(ft, "features", function(feature) feature / sum(feature))

# Map over samples
map(ft, "samples", function(sample) sample / sum(sample))

}
\seealso{
Other apply functions: 
\code{\link{map_with_index}()},
\code{\link{map_with_name}()}
}
\concept{apply functions}
