% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{map_features}
\alias{map_features}
\title{Apply a function over features of a FeatureTable}
\usage{
map_features(ft, fn, ...)
}
\arguments{
\item{ft}{A FeatureTable object.}

\item{fn}{The function to be applied over the features.}

\item{...}{Optional arguments to \code{fn} (if \code{fn} is a function).}
}
\value{
A new FeatureTable with the result of applying \code{fn} over features.
}
\description{
Returns a new FeatureTable obtained by applying a function to features of the \code{$data} field of a \code{FeatureTable}.
}
\seealso{
Other mapping functions: 
\code{\link{map_features_with_index}()},
\code{\link{map_features_with_name}()},
\code{\link{map_samples_with_index}()},
\code{\link{map_samples_with_name}()},
\code{\link{map_samples}()},
\code{\link{map_with_index}()},
\code{\link{map_with_name}()},
\code{\link{map}()}
}
\concept{mapping functions}
