% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_with_index}
\alias{map_with_index}
\alias{map_features_with_index}
\alias{map_samples_with_index}
\title{Apply a function over FeatureTable margins and return a new FeatureTable.}
\usage{
map_with_index(ft, margin, fn, ...)

map_features_with_index(ft, fn, ...)

map_samples_with_index(ft, fn, ...)
}
\arguments{
\item{ft}{A FeatureTable.}

\item{margin}{Margin to apply the function over.  E.g., \code{1} or
\code{"samples"} indicates rows, \code{2} or \code{"features"} indicates
columns.}

\item{fn}{The function to be applied over the specified margin.  This
function should have at least 2 parameters, the data to be applied over,
and the index of the current row or column.  See details.}

\item{...}{Optional arguments to \code{fn}.}
}
\value{
A new FeatureTable with the result of applying \code{fn} over
  \code{margin}.
}
\description{
Returns a new FeatureTable by applying a function to margins of the
\code{$data} field of a \code{FeatureTable}.
}
\details{
This function is similar to \code{\link{map}}, except that the function
should accept index of the row or column currently being applied over.  See
Examples.

Note that if you want to include something from the \code{feature_data} or
\code{sample_data} in the mapping function, you probably want to use
\code{\link{map_with_name}}.

Named helpers can also be used rather than specifying the \code{margin}
argument.

\itemize{
  \item \code{\link{map_features_with_index}}
  \item \code{\link{map_samples_with_index}}
}
}
\examples{
data(ft)

# Make up some fake lengths for the features.
lengths <- 1:ft$num_features()

# Map over features using the index to divide the feature by its length.
ft$map_with_index("features", function(feature, index) feature / lengths[[index]])

# And of course, the s3 function is also available!
map_with_index(ft, "features", function(feature, index) feature / lengths[[index]])

}
\seealso{
Other apply functions: 
\code{\link{map_with_name}()},
\code{\link{map}()}
}
\concept{apply functions}
