% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{map_with_name}
\alias{map_with_name}
\title{Apply a function over FeatureTable margins.}
\usage{
map_with_name(ft, margin, fn, ...)
}
\arguments{
\item{ft}{A FeatureTable object.}

\item{margin}{Which margin the function will be applied over.  E.g., \code{1} or \code{"samples"} indicates rows, \code{2} or \code{"features"} indicates columns.}

\item{fn}{The function to be applied over the specified margin.  This function should have at least 2 parameters, the data applied over, and the name of the current row or column.}

\item{...}{Optional arguments to \code{fn} (if \code{fn} is a function).}
}
\value{
A new FeatureTable with the result of applying \code{fn} over \code{margin}.
}
\description{
Returns a new FeatureTable obtained by applying a function to margins of the \code{$data} field of a \code{FeatureTable}.
}
\seealso{
Other mapping functions: 
\code{\link{map_features_with_index}()},
\code{\link{map_features_with_name}()},
\code{\link{map_features}()},
\code{\link{map_samples_with_index}()},
\code{\link{map_samples_with_name}()},
\code{\link{map_samples}()},
\code{\link{map_with_index}()},
\code{\link{map}()}
}
\concept{mapping functions}
