% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{max.FeatureTable}
\alias{max.FeatureTable}
\alias{min.FeatureTable}
\alias{non_zero_min}
\alias{non_zero_min.FeatureTable}
\alias{size}
\alias{size.FeatureTable}
\title{Minimum, maximum, size, etc., of FeatureTable data field.}
\usage{
\method{max}{FeatureTable}(ft, ...)

\method{min}{FeatureTable}(ft, ...)

non_zero_min(ft, ...)

\method{non_zero_min}{FeatureTable}(ft, ...)

size(ft, ...)

\method{size}{FeatureTable}(ft)
}
\value{
The min (\code{min}), non-zero min (\code{non_zero_min}), and max (\code{max}) values.  \code{size} returns the number of elements in the \code{data} field (i.e., \code{nrow * ncol} for the \code{data} field).
}
\description{
Get the minimum, maximum, size, and other features of the FeatureTable \code{data} field.
}
\details{
\code{non_zero_min} returns the smallest non-zero number in the \code{data} field.  If any of the values in the \code{data} field are less than zero, then a warning will be given.  In this case, you probably don't want to use \code{non_zero_min}.
}
\examples{
ft <- FeatureTable$new(matrix(0:8, 3, 3))
stopifnot(min(ft) == 0)
stopifnot(non_zero_min(ft) == 1)
stopifnot(max(ft) == 8)
stopifnot(size(ft) == 9)

}
\seealso{
[featuretable::is_count_table()]
}
