% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{nsamples}
\alias{nsamples}
\title{The number of samples/observations/rows of a FeatureTable}
\usage{
nsamples(ft)
}
\arguments{
\item{ft}{A FeatureTable object.}
}
\value{
The number of rows in the FeatureTable.
}
\description{
Return the number of samples/observations/rows in the FeatureTable.
}
\seealso{
Other dimension functions: 
\code{\link{dim.FeatureTable}()},
\code{\link{nfeatures}()},
\code{\link{nobservations}()},
\code{\link{num_features}()},
\code{\link{num_observations}()},
\code{\link{num_samples}()}
}
\concept{dimension functions}
