% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{num_features}
\alias{num_features}
\alias{num_features.FeatureTable}
\alias{nfeatures}
\alias{nfeatures.FeatureTable}
\title{The number of features/columns of a FeatureTable}
\usage{
num_features(ft)

\method{num_features}{FeatureTable}(ft)

nfeatures(ft)

\method{nfeatures}{FeatureTable}(ft)
}
\arguments{
\item{ft}{A FeatureTable object.}
}
\value{
The number of features/columns in the FeatureTable.
}
\description{
Return the number of features/columns in the FeatureTable.
}
\seealso{
Other dimension functions: 
\code{\link{num_samples}()}
}
\concept{dimension functions}
