% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{observation_names}
\alias{observation_names}
\title{Samples/observations/rows of a FeatureTable}
\usage{
observation_names(ft)
}
\arguments{
\item{ft}{A FeatureTable object.}
}
\value{
The names of samples/observations/rows in the FeatureTable.
}
\description{
Return the names of samples/observations/rows in the FeatureTable.
}
\seealso{
Other row name functions: 
\code{\link{sample_names}()}
}
\concept{row name functions}
