% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pca_biplot}
\alias{pca_biplot}
\alias{pca_biplot.FeatureTable}
\title{PCA Biplot}
\usage{
pca_biplot(ft, use_biplotr = FALSE, include_sample_data = FALSE, ...)

\method{pca_biplot}{FeatureTable}(ft, use_biplotr = FALSE, include_sample_data = FALSE, ...)
}
\arguments{
\item{ft}{A FeatureTable}

\item{use_biplotr}{Use biplotR for nice biplots.  Must be installed.}

\item{incude_sample_data}{Do you want to include sample data?  If \code{TRUE},
then you will have access to lots of nice sample data to help annotate your
plots.  See Examples.}
}
\description{
PCA Biplot
}
\details{
If installed, you can use \href{https://github.com/mooreryan/biplotr}{biplotR}
for nice looking biplots!  Be sure that if you use_biplotr = TRUE, not to set
options for the base::biplot function.

See documentation for \code{?biplotr::pca_biplot} for more info.

If you don't want to use \href{https://github.com/mooreryan/biplotr}{biplotR},
then this function is a wrapper for \code{base::biplot}.
}
\examples{
# TODO!

}
