% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{plot.FeatureTable}
\alias{plot.FeatureTable}
\title{FeatureTable summary plots}
\usage{
\method{plot}{FeatureTable}(ft, ...)
}
\arguments{
\item{ft}{A FeatureTable}

\item{...}{Extra args}

\item{num_features}{How many features to show on the plot? The rest of the
features get put into an "Other" category. "Other" will always be gray if
you use one of the built in palettes.}

\item{fill}{If FALSE, don't show the features, if anything else, show the
features.}

\item{palette}{Which palette to use? ("kelly", "muted", "bright", "vibrant",
"high contrast").  If there are more features than colors in the palette,
 the colors will be recycled.}

\item{show_legend}{Show the legend on the right side?}

\item{legend_title}{Custom legend title}

\item{plot_title}{Custom plot title}

\item{xlab}{Custom xlab}

\item{ylab}{Custom ylab}

\item{axis.text.x}{Custom \code{axis.text.x}.  Should be an
\code{element_text}.}
}
\value{
A `ggplot2` object.
}
\description{
FeatureTable summary plots
}
