% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_abundance}
\alias{relative_abundance}
\title{Calculate relative abundance (proportions).}
\usage{
relative_abundance(sample, multiplier = 1)
}
\arguments{
\item{sample}{A vector}

\item{multiplier}{Multiply the result by this.  E.g., \code{multiplier = 100}
would convert to percentages.}
}
\value{
"TODO"
}
\description{
Calculate relative abundance (proportions).
}
\examples{
"TODO"

}
