% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{replace_zeros.FeatureTable}
\alias{replace_zeros.FeatureTable}
\title{Replacing zeros.}
\usage{
\method{replace_zeros}{FeatureTable}(
  ft,
  replacement = 0.05,
  tol = .Machine$double.eps^0.5,
  use_cmultRepl = FALSE,
  ...
)
}
\arguments{
\item{ft}{A FeatureTable}

\item{replacement}{(Ignored if \code{use_cmultRepl = TRUE})}

\item{tol}{(Ignored if \code{use_cmultRepl = TRUE})}

\item{use_cmultRepl}{TRUE/FALSE whether to use \code{cmultRepl} function.}

\item{...}{Extra arguments (i.g., passed to \code{cmultRepl})}
}
\value{
A new FeatureTable with the zeros in \code{data} replaced.
}
\description{
Replacing zeros.
}
