% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{replace_zeros}
\alias{replace_zeros}
\alias{replace_zeros.FeatureTable}
\title{Replacing zeros.}
\usage{
replace_zeros(ft, ...)

\method{replace_zeros}{FeatureTable}(
  ft,
  replacement = 0.05,
  tol = .Machine$double.eps^0.5,
  use_cmultRepl = FALSE,
  ...
)
}
\arguments{
\item{ft}{A FeatureTable}

\item{...}{Extra arguments (i.g., passed to \code{cmultRepl})}

\item{replacement}{(Ignored if \code{use_cmultRepl = TRUE})}

\item{tol}{(Ignored if \code{use_cmultRepl = TRUE})}

\item{use_cmultRepl}{TRUE/FALSE whether to use \code{cmultRepl} function.}
}
\value{
A new FeatureTable with the zeros in \code{data} replaced.
}
\description{
Replacing zeros.
}
\details{
If \code{use_cmultRepl = TRUE} but \code{zCompositions::cmultRepl} is not
availabel, it will raise an error.
}
\seealso{
\code{\link{cmultRepl}}
}
