% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{shared_features}
\alias{shared_features}
\alias{shared_features.FeatureTable}
\alias{keep_shared_features}
\alias{keep_shared_features.FeatureTable}
\alias{shared_feature_names}
\alias{shared_feature_names.FeatureTable}
\title{Merging FeatureTables by shared features.}
\usage{
shared_features(ft, method, other)

\method{shared_features}{FeatureTable}(ft, method, other)

keep_shared_features(ft, other)

\method{keep_shared_features}{FeatureTable}(ft, other)

shared_feature_names(ft, other)

\method{shared_feature_names}{FeatureTable}(ft, other)
}
\arguments{
\item{ft}{A FeatureTable}

\item{method}{One of "keep" or "names".}

\item{other}{Another FeatureTable}
}
\value{
If \code{method = "keep"}, return a new FeatureTable containing only features
shared between \code{ft} (or \code{self} for the R6 method) and \code{other}.

If \code{method = "names"}, return the shared feature names between \code{ft}
(or \code{self} for the R6 method) and \code{other}.
}
\description{
Merge FeatureTables by keeping only features that are shared between them
(base on feature names).
}
\details{
If you just have a list of feature names, you don't need these, you'd use
\code{\link{keep}} directly.

Consider this API pretty unstable.  I haven't decided the on how I want these
functions to behave in the end.
}
