% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Data used for examples and to test the FeatureTable functions.}
\format{
A list
\describe{
  \item{nsamples}{the number of samples}
  \item{nfeatures}{the number of features}
  \item{count_table}{a matrix with counts}
  \item{feature_data}{a data frame with info about features}
  \item{sample_data}{a data frame with info about samples}
  \item{expected_feature_data}{what the feature data looks like after making the FeatureTable}
  \item{expected_sample_data}{what the samples data looks like after making the FeatureTable}
}
}
\usage{
testdata
}
\description{
A small dataset for testing and making examples for FeatureTable
documentation.
}
\keyword{datasets}
