% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_exporting.R
\name{as.phyloseq}
\alias{as.phyloseq}
\title{Convert FeatureTable to phyloseq object.}
\usage{
as.phyloseq(ft)
}
\arguments{
\item{ft}{A FeatureTable}
}
\value{
a phyloseq object
}
\description{
If the 'phyloseq' package is not installed, it raises an Error.
}
