% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{dim.FeatureTable}
\alias{dim.FeatureTable}
\title{Dimensions of FeatureTable data.}
\usage{
\method{dim}{FeatureTable}(x)
}
\arguments{
\item{x}{A FeatureTable object.}
}
\value{
The dimension of the feature_table.
}
\description{
Return the dimensions of the FeatureTable \code{data} field.
}
\examples{
ft <- FeatureTable$new(matrix(1:12, 3, 4))
stopifnot(ft$dim() == c(3, 4))
stopifnot(nrow(ft) == 3)
stopifnot(ncol(ft) == 4)

}
\seealso{
\code{\link{dim}} for generic method.
}
\concept{general s3-like functions}
