% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{feature_names}
\alias{feature_names}
\title{Features/columns of a FeatureTable}
\usage{
feature_names(ft)
}
\arguments{
\item{ft}{A FeatureTable object.}
}
\value{
The names of features/columns in the FeatureTable.
}
\description{
Return the names of features/columns in the FeatureTable.
}
