% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuretable_package.R
\docType{package}
\name{featuretable-package}
\alias{featuretable}
\alias{featuretable-package}
\title{FeatureTable}
\description{
A FeatureTable R6 and s3 class for managing feature tables and data associated with its features and observations.
}
\keyword{internal}
