% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table.R
\name{FeatureTable}
\alias{FeatureTable}
\title{FeatureTable}
\description{
R6 Class representing a feature table and its associated data.
}
\details{
A FeatureTable has a feature_table, sample_data, and feature_data.
}
\examples{

## ------------------------------------------------
## Method `FeatureTable$new`
## ------------------------------------------------

"TODO"
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureTable$new()}}
\item \href{#method-dim}{\code{FeatureTable$dim()}}
\item \href{#method-num_samples}{\code{FeatureTable$num_samples()}}
\item \href{#method-nsamples}{\code{FeatureTable$nsamples()}}
\item \href{#method-num_observations}{\code{FeatureTable$num_observations()}}
\item \href{#method-nobservations}{\code{FeatureTable$nobservations()}}
\item \href{#method-nrow}{\code{FeatureTable$nrow()}}
\item \href{#method-num_features}{\code{FeatureTable$num_features()}}
\item \href{#method-nfeatures}{\code{FeatureTable$nfeatures()}}
\item \href{#method-ncol}{\code{FeatureTable$ncol()}}
\item \href{#method-sample_names}{\code{FeatureTable$sample_names()}}
\item \href{#method-observation_names}{\code{FeatureTable$observation_names()}}
\item \href{#method-feature_names}{\code{FeatureTable$feature_names()}}
\item \href{#method-print}{\code{FeatureTable$print()}}
\item \href{#method-apply}{\code{FeatureTable$apply()}}
\item \href{#method-apply_with_index}{\code{FeatureTable$apply_with_index()}}
\item \href{#method-apply_with_name}{\code{FeatureTable$apply_with_name()}}
\item \href{#method-apply_features}{\code{FeatureTable$apply_features()}}
\item \href{#method-apply_features_with_index}{\code{FeatureTable$apply_features_with_index()}}
\item \href{#method-apply_features_with_name}{\code{FeatureTable$apply_features_with_name()}}
\item \href{#method-apply_samples}{\code{FeatureTable$apply_samples()}}
\item \href{#method-apply_samples_with_index}{\code{FeatureTable$apply_samples_with_index()}}
\item \href{#method-apply_samples_with_name}{\code{FeatureTable$apply_samples_with_name()}}
\item \href{#method-reduce}{\code{FeatureTable$reduce()}}
\item \href{#method-reduce_features}{\code{FeatureTable$reduce_features()}}
\item \href{#method-reduce_samples}{\code{FeatureTable$reduce_samples()}}
\item \href{#method-reduce_all}{\code{FeatureTable$reduce_all()}}
\item \href{#method-map}{\code{FeatureTable$map()}}
\item \href{#method-map_with_index}{\code{FeatureTable$map_with_index()}}
\item \href{#method-map_with_name}{\code{FeatureTable$map_with_name()}}
\item \href{#method-map_features}{\code{FeatureTable$map_features()}}
\item \href{#method-map_features_with_index}{\code{FeatureTable$map_features_with_index()}}
\item \href{#method-map_features_with_name}{\code{FeatureTable$map_features_with_name()}}
\item \href{#method-map_samples}{\code{FeatureTable$map_samples()}}
\item \href{#method-map_samples_with_index}{\code{FeatureTable$map_samples_with_index()}}
\item \href{#method-map_samples_with_name}{\code{FeatureTable$map_samples_with_name()}}
\item \href{#method-keep}{\code{FeatureTable$keep()}}
\item \href{#method-keep_features}{\code{FeatureTable$keep_features()}}
\item \href{#method-keep_samples}{\code{FeatureTable$keep_samples()}}
\item \href{#method-shared_features}{\code{FeatureTable$shared_features()}}
\item \href{#method-keep_shared_features}{\code{FeatureTable$keep_shared_features()}}
\item \href{#method-shared_feature_names}{\code{FeatureTable$shared_feature_names()}}
\item \href{#method-as_phyloseq}{\code{FeatureTable$as_phyloseq()}}
\item \href{#method-is_count_table}{\code{FeatureTable$is_count_table()}}
\item \href{#method-replace_zeros}{\code{FeatureTable$replace_zeros()}}
\item \href{#method-clr}{\code{FeatureTable$clr()}}
\item \href{#method-pca_biplot}{\code{FeatureTable$pca_biplot()}}
\item \href{#method-core_microbiome}{\code{FeatureTable$core_microbiome()}}
\item \href{#method-collapse}{\code{FeatureTable$collapse()}}
\item \href{#method-collapse_features}{\code{FeatureTable$collapse_features()}}
\item \href{#method-collapse_samples}{\code{FeatureTable$collapse_samples()}}
\item \href{#method-max}{\code{FeatureTable$max()}}
\item \href{#method-min}{\code{FeatureTable$min()}}
\item \href{#method-non_zero_min}{\code{FeatureTable$non_zero_min()}}
\item \href{#method-size}{\code{FeatureTable$size()}}
\item \href{#method-plot}{\code{FeatureTable$plot()}}
\item \href{#method-clone}{\code{FeatureTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new FeatureTable object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$new(
  feature_table,
  feature_data = NULL,
  sample_data = NULL,
  feature_table_rows_are_samples = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_table}}{Feature table, count table, or something similar.}

\item{\code{feature_data}}{Feature data (data about the features).}

\item{\code{sample_data}}{Sample/observation data (data about the samples/observations).}

\item{\code{feature_table_rows_are_samples}}{Are the rows of the feature_table samples/observations?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FeatureTable` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{"TODO"
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\if{latex}{\out{\hypertarget{method-dim}{}}}
\subsection{Method \code{dim()}}{
Return the dimensions of the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dimension of the feature_table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_samples"></a>}}
\if{latex}{\out{\hypertarget{method-num_samples}{}}}
\subsection{Method \code{num_samples()}}{
Return the number of samples/observations/rows in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$num_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of rows in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nsamples"></a>}}
\if{latex}{\out{\hypertarget{method-nsamples}{}}}
\subsection{Method \code{nsamples()}}{
Return the number of samples/observations/rows in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$nsamples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of rows in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_observations"></a>}}
\if{latex}{\out{\hypertarget{method-num_observations}{}}}
\subsection{Method \code{num_observations()}}{
Return the number of samples/observations/rows in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$num_observations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of rows in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nobservations"></a>}}
\if{latex}{\out{\hypertarget{method-nobservations}{}}}
\subsection{Method \code{nobservations()}}{
Return the number of samples/observations/rows in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$nobservations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of rows in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nrow"></a>}}
\if{latex}{\out{\hypertarget{method-nrow}{}}}
\subsection{Method \code{nrow()}}{
Return the number of samples/observations/rows in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$nrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of rows in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_features"></a>}}
\if{latex}{\out{\hypertarget{method-num_features}{}}}
\subsection{Method \code{num_features()}}{
Return the number of features/columns in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$num_features()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of columns in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nfeatures"></a>}}
\if{latex}{\out{\hypertarget{method-nfeatures}{}}}
\subsection{Method \code{nfeatures()}}{
Return the number of features/columns in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$nfeatures()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of columns in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ncol"></a>}}
\if{latex}{\out{\hypertarget{method-ncol}{}}}
\subsection{Method \code{ncol()}}{
Return the number of features/columns in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$ncol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of columns in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_names"></a>}}
\if{latex}{\out{\hypertarget{method-sample_names}{}}}
\subsection{Method \code{sample_names()}}{
Return the names of samples/observations/rows in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$sample_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The names of samples/observations/rows in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-observation_names"></a>}}
\if{latex}{\out{\hypertarget{method-observation_names}{}}}
\subsection{Method \code{observation_names()}}{
Return the names of samples/observations/rows in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$observation_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The names of samples/observations/rows in the FeatureTable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feature_names"></a>}}
\if{latex}{\out{\hypertarget{method-feature_names}{}}}
\subsection{Method \code{feature_names()}}{
Return the names of features/columns in the FeatureTable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$feature_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The names of features/columns in the FeatureTable.
Convert FeatureTable to phyloseq object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply"></a>}}
\if{latex}{\out{\hypertarget{method-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply(margin, fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_with_index"></a>}}
\if{latex}{\out{\hypertarget{method-apply_with_index}{}}}
\subsection{Method \code{apply_with_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_with_index(margin, fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_with_name"></a>}}
\if{latex}{\out{\hypertarget{method-apply_with_name}{}}}
\subsection{Method \code{apply_with_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_with_name(margin, fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_features"></a>}}
\if{latex}{\out{\hypertarget{method-apply_features}{}}}
\subsection{Method \code{apply_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_features(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_features_with_index"></a>}}
\if{latex}{\out{\hypertarget{method-apply_features_with_index}{}}}
\subsection{Method \code{apply_features_with_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_features_with_index(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_features_with_name"></a>}}
\if{latex}{\out{\hypertarget{method-apply_features_with_name}{}}}
\subsection{Method \code{apply_features_with_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_features_with_name(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_samples"></a>}}
\if{latex}{\out{\hypertarget{method-apply_samples}{}}}
\subsection{Method \code{apply_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_samples(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_samples_with_index"></a>}}
\if{latex}{\out{\hypertarget{method-apply_samples_with_index}{}}}
\subsection{Method \code{apply_samples_with_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_samples_with_index(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_samples_with_name"></a>}}
\if{latex}{\out{\hypertarget{method-apply_samples_with_name}{}}}
\subsection{Method \code{apply_samples_with_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$apply_samples_with_name(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reduce"></a>}}
\if{latex}{\out{\hypertarget{method-reduce}{}}}
\subsection{Method \code{reduce()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$reduce(margin, fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reduce_features"></a>}}
\if{latex}{\out{\hypertarget{method-reduce_features}{}}}
\subsection{Method \code{reduce_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$reduce_features(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reduce_samples"></a>}}
\if{latex}{\out{\hypertarget{method-reduce_samples}{}}}
\subsection{Method \code{reduce_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$reduce_samples(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reduce_all"></a>}}
\if{latex}{\out{\hypertarget{method-reduce_all}{}}}
\subsection{Method \code{reduce_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$reduce_all(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\if{latex}{\out{\hypertarget{method-map}{}}}
\subsection{Method \code{map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map(margin, fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_with_index"></a>}}
\if{latex}{\out{\hypertarget{method-map_with_index}{}}}
\subsection{Method \code{map_with_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_with_index(margin, fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_with_name"></a>}}
\if{latex}{\out{\hypertarget{method-map_with_name}{}}}
\subsection{Method \code{map_with_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_with_name(margin, fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_features"></a>}}
\if{latex}{\out{\hypertarget{method-map_features}{}}}
\subsection{Method \code{map_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_features(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_features_with_index"></a>}}
\if{latex}{\out{\hypertarget{method-map_features_with_index}{}}}
\subsection{Method \code{map_features_with_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_features_with_index(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_features_with_name"></a>}}
\if{latex}{\out{\hypertarget{method-map_features_with_name}{}}}
\subsection{Method \code{map_features_with_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_features_with_name(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_samples"></a>}}
\if{latex}{\out{\hypertarget{method-map_samples}{}}}
\subsection{Method \code{map_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_samples(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_samples_with_index"></a>}}
\if{latex}{\out{\hypertarget{method-map_samples_with_index}{}}}
\subsection{Method \code{map_samples_with_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_samples_with_index(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_samples_with_name"></a>}}
\if{latex}{\out{\hypertarget{method-map_samples_with_name}{}}}
\subsection{Method \code{map_samples_with_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$map_samples_with_name(fn, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep"></a>}}
\if{latex}{\out{\hypertarget{method-keep}{}}}
\subsection{Method \code{keep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$keep(margin, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_features"></a>}}
\if{latex}{\out{\hypertarget{method-keep_features}{}}}
\subsection{Method \code{keep_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$keep_features(predicate, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_samples"></a>}}
\if{latex}{\out{\hypertarget{method-keep_samples}{}}}
\subsection{Method \code{keep_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$keep_samples(predicate, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shared_features"></a>}}
\if{latex}{\out{\hypertarget{method-shared_features}{}}}
\subsection{Method \code{shared_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$shared_features(method, other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_shared_features"></a>}}
\if{latex}{\out{\hypertarget{method-keep_shared_features}{}}}
\subsection{Method \code{keep_shared_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$keep_shared_features(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shared_feature_names"></a>}}
\if{latex}{\out{\hypertarget{method-shared_feature_names}{}}}
\subsection{Method \code{shared_feature_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$shared_feature_names(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_phyloseq"></a>}}
\if{latex}{\out{\hypertarget{method-as_phyloseq}{}}}
\subsection{Method \code{as_phyloseq()}}{
If the 'phyloseq' package is not installed, it raises an Error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$as_phyloseq()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ft}}{A FeatureTable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a phyloseq object
Is the FeatureTable a count table?
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_count_table"></a>}}
\if{latex}{\out{\hypertarget{method-is_count_table}{}}}
\subsection{Method \code{is_count_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$is_count_table()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ft}}{A FeatureTable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if \code{feature_table} contains only natural numbers (counting numbers >= 0), FALSE otherwise.
Replacing zeros.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_zeros"></a>}}
\if{latex}{\out{\hypertarget{method-replace_zeros}{}}}
\subsection{Method \code{replace_zeros()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$replace_zeros(
  replacement = 0.05,
  tol = .Machine$double.eps^0.5,
  use_cmultRepl = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replacement}}{(Ignored if \code{use_cmultRepl = TRUE})}

\item{\code{tol}}{(Ignored if \code{use_cmultRepl = TRUE})}

\item{\code{use_cmultRepl}}{TRUE/FALSE whether to use \code{cmultRepl} function.}

\item{\code{...}}{Extra arguments (i.g., passed to \code{cmultRepl})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new FeatureTable with the zeros in \code{data} replaced.
Centered log ratio.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clr"></a>}}
\if{latex}{\out{\hypertarget{method-clr}{}}}
\subsection{Method \code{clr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$clr(base = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base}}{Base of logarithm.}

\item{\code{ft}}{A Feature table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A FeatureTable with the \code{data} transformed with centered log ratio.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pca_biplot"></a>}}
\if{latex}{\out{\hypertarget{method-pca_biplot}{}}}
\subsection{Method \code{pca_biplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$pca_biplot(use_biplotr = FALSE, include_sample_data = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-core_microbiome"></a>}}
\if{latex}{\out{\hypertarget{method-core_microbiome}{}}}
\subsection{Method \code{core_microbiome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$core_microbiome(
  detection_limit = 1,
  min_sample_proportion = NULL,
  max_sample_proportion = NULL,
  min_samples = NULL,
  max_samples = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collapse"></a>}}
\if{latex}{\out{\hypertarget{method-collapse}{}}}
\subsection{Method \code{collapse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$collapse(margin, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collapse_features"></a>}}
\if{latex}{\out{\hypertarget{method-collapse_features}{}}}
\subsection{Method \code{collapse_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$collapse_features(by, keep_na = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collapse_samples"></a>}}
\if{latex}{\out{\hypertarget{method-collapse_samples}{}}}
\subsection{Method \code{collapse_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$collapse_samples(by, keep_na = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-max"></a>}}
\if{latex}{\out{\hypertarget{method-max}{}}}
\subsection{Method \code{max()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$max(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-min"></a>}}
\if{latex}{\out{\hypertarget{method-min}{}}}
\subsection{Method \code{min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$min(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-non_zero_min"></a>}}
\if{latex}{\out{\hypertarget{method-non_zero_min}{}}}
\subsection{Method \code{non_zero_min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$non_zero_min(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$size(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$plot(
  num_features = 8,
  other_feature_name = "Other",
  fill = TRUE,
  palette = "kelly",
  show_legend = TRUE,
  legend_title = NULL,
  plot_title = NULL,
  xlab = NULL,
  ylab = NULL,
  axis.text.x = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
