% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_s3.R
\name{num_features}
\alias{num_features}
\title{The number of features/columns of a FeatureTable}
\usage{
num_features(ft)
}
\arguments{
\item{ft}{A FeatureTable object.}
}
\value{
The number of columns in the FeatureTable.
}
\description{
Return the number of features/columns in the FeatureTable.
}
\seealso{
Other dimension functions: 
\code{\link{dim.FeatureTable}()},
\code{\link{nfeatures}()},
\code{\link{nobservations}()},
\code{\link{nsamples}()},
\code{\link{num_observations}()},
\code{\link{num_samples}()}
}
\concept{dimension functions}
